lab3/requirements.txt
numpy>=1.24,<2.1
h5py>=3.10
matplotlib>=3.8
faiss-cpu==1.8.0.post1
